Move the breakpoint script to a separate file
